+++ packages/db-drizzlepg/src/schema/auth/users.ts
@@ -1,11 +1,9 @@
-import { relations } from 'drizzle-orm'
 import { text } from 'drizzle-orm/pg-core'
 
 import { citext } from '../custom-types.js'
 import { namedUnique, withSurrogateId } from '../helpers.js'
 
 import { authSchema } from './schema.js'
-import { sessionsTable } from './sessions.js'
 
 export const usersTable = authSchema.table(
   'users',
@@ -16,10 +14,6 @@ export const usersTable = authSchema.table(
   },
   (t) => [namedUnique(t.username)],
 )
-
-export const userRelations = relations(usersTable, ({ many }) => ({
-  sessions: many(sessionsTable),
-}))
 
 export type UserEntity = Omit<typeof usersTable.$inferSelect, 'hashedPassword'>
 export type UserEntityInsert = typeof usersTable.$inferInsert
